cleanup(angular): add e2e for mf ssr
